+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Update FLTK and cfltk.
 - cleanup fltk.patch.
 - Fix broken image::Pixmap.
+- Add asserts for supported window icon images.
 
 ## [0.14.11] - 2021-02-17
 ### Changes